requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
sentence-transformers>=3.0
faiss-cpu>=1.8
numpy>=1.26
pytest>=8.2
